chore: move namespace pkg under testutil dir
